pyproject.toml
[tool.poetry]
name = "umami-app"
version = "0.1.0"
description = ""
authors = ["Agustin <agustin>"]

[tool.poetry.dependencies]
python = "^3.6"
django = "^3.0.2"
djangorestframework = "^3.11.0"
whitenoise = "^5.0.1"
gunicorn = "^20.0.4"
django-heroku = "^0.3.1"
django-cors-headers = "^3.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
